fix(cli): add missing stdint.h include for uintptr_t
Linux builds failed with 'uintptr_t undeclared' because stdint.h was
not included. Also cast utf8_len to size_t to silence the remaining
-Wsign-conversion warning.

+++ src/cli.c
@@ -2,6 +2,7 @@
 
 #include <ctype.h>
 #include <errno.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -58,7 +59,7 @@ static char *win32_utf8(const char *ansi) {
         free(wide);
         return NULL;
     }
-    utf8 = (char *)malloc(utf8_len);
+    utf8 = (char *)malloc((size_t)utf8_len);
     if (utf8 == NULL) {
         free(wide);
         return NULL;
